website+ui: Move list of courses into one place
There are several issues open for courses missing in different places, and the code seems duplicated and inconsistent about handling these without good reason.

This might not be the right list of courses to be going ahead with in the near future, but at least this gives us one place where we can update it.

Maybe in the future this moves to TinaCMS, or links in with the platform directly, so this isn't necessary. But this is a step closer to having less chaos about what courses we list.