room: Prevent API recursion depth errors
On each request, we were modifying the requestBody schema. Because this was captured in the higher level closure it was being nested, until eventually checking it was no longer feasible and we got errors.

This fixes the problem by not modifying the opts object. Should have listened to the ESLint warning in the first place 🤦