libraries/ui: Add streaming response support to makeMakeApiRoute
This just disables response schema checking if the response is supposed to be a streaming response. Users of makeMakeApi route are expected to use the special 'StreamingResponseSchema' response body to enable this.

This is primarily for course-demos, where we're likely to want more AI streaming demos. Being able to use makeMakeApiRoute for these gives us things like authentication, request validation, nice typings, and error handling "for free", which cleans up the API code considerably.

In future we could try to make this even nicer - but it's already a good improvement.