devex: Fix package.json for eslint-config linking to eslint-plugin-custom
This was causing mysterious npm errors when installing packages, but only sometimes, such as:

```
npm error Cannot read properties of undefined (reading 'extraneous')
```

The package name for our custom eslint plugin was wrong in eslint-config's package.json. I figured this out by editing the NPM source code to print out more details of the error, and it was consistently when resolving the @bluedot/eslint-config package. I'm fairly certain this was the cause of the problem, but I can't be certain - we'll see if the bug reappears.