devex: Make setup of .env.local easier (#719)
Most apps create `.env.local` automatically if it doesn't exist on `postinstall`, but not all. This PR makes it consistent across all apps.

Also the Slack token isn't really required - shoving 'IGNORE_SLACK_ALERTS' as the API token will generally cause it to fail quietly which is fine locally. I've filled it in like this to make getting started easier.