devex: Fix deployment `npm ci`
Narrowing it down to the package fails because npm tries to run postinstall in all the other packages, which can fail if they don't have their dependencies installed.

Ideally we could not run these postinstalls. But npm doesn't support this. (the closest is the `ignore-scripts` option, and then doing postinstall manually for just this workspace, but this will mess with dependency scripts too: https://stackoverflow.com/questions/23505318/npm-disable-postinstall-script-for-package )