Serve fonts from bluedot.org consistently (#770)
This reduces duplication in the repo, and makes it easier to use the ui library wherever because the right fonts are always pulled in (i.e. it doesn't depend on whether the font is installed in your particular app).

I've added some .gitkeeps to the public folders just so it's still easy for people to add static assets. (see https://www.freecodecamp.org/news/what-is-gitkeep/ for more details)