[Feat] UX fixes on chunks (#1229)
* Remove fixed height constraints from sidebar chunk items

Remove min-h-[100px] and min-h-[68px] classes from sidebar chunk buttons and content containers to allow dynamic height based on content length.

* Remove UnitFeedback component and update UnitLayout

Remove the UnitFeedback component along with its tests and snapshots, and update UnitLayout to no longer include the feedback functionality.

We're not using this functionality, so we don't need it on the page. It adds clutter. If we want to add unit feedback back in the future, we can do so then.