`AnnouncementBanner` should be a pure client component
Since the component needs to read/write local storage state (i.e. using
Zustand) we need to tell Next that this component should be rendered
purely on the client.

Only using `use client` still leads to pre-rendering on the server,
which was causing a hydration issue. We need to import the component in
a `dynamic` call to ensure we properly skip SSR:
https://nextjs.org/docs/app/guides/lazy-loading#skipping-ssr