Use number of calendar days when determining relative date time (#1561)
Prefer number of calendar days to number of hours

Previously, we would base the number of days off of the number of hours
between the dates (rounded down). However, this creates confusion -
imagine it's 10pm, and your discussion is in 1 day 12 hours. Previously,
we would say your discussion is in 1 day (since we rounded down). Now,
we calculate the number of calendar days (in this case, 2).