Deduplicate response by `unitResourceIdRead`
In the unlikely event of having multiple resource completions for a user
for a resource, we would fetch too many responses, leading to something
like '3 of 2 completed'.

It would be better to enforce uniqueness at a database level, but since
this doesn't seem possible in Airtable I'm adding this as a defensive
layer.